Made basic trajectory

+++ auv_guidance/include/auv_guidance/segment_planner.hpp
@@ -15,10 +15,10 @@ namespace AUVGuidance
 class SegmentPlanner
 {
   private:
-    float distance_, cruiseSpeed_, acceleration_;
-    float cruiseDuration_, initialSpeed_, maxSpeed_, finalSpeed_;
+    double distance_, cruiseSpeed_, acceleration_;
+    double cruiseDuration_, initialSpeed_, maxSpeed_, finalSpeed_;
     int accelSeq_;
-    float t1_, t2_, tMid_, tEnd_; // Key times. At cruiseSpeed for in the time interval [t1_, t2_]
+    double t1_, t2_, tMid_, tEnd_; // Key times. At cruiseSpeed for in the time interval [t1_, t2_]
     bool accelerate_;
 
   public:
@@ -26,12 +26,12 @@ class SegmentPlanner
     static const int SEQ_START = 1;
     static const int SEQ_END = 2;
     static const int SEQ_BOTH = 3;
-    static const float DEFAULT_SPEED = 1.0;
+    static const double DEFAULT_SPEED = 1.0;
 
-    SegmentPlanner(float distance, float nominalSpeed, float accel = 0.0, int seq = SegmentPlanner::SEQ_NONE);
+    SegmentPlanner(double distance, double nominalSpeed, double accel = 0.0, int seq = SegmentPlanner::SEQ_NONE);
     void initMotionPlanner();
-    float getTravelTime();
-    Vector2f computeState(float t);
+    double getTravelTime();
+    Vector2d computeState(double t);
 };
 } // namespace AUVGuidance
